Emit console.trace instead of throwing errors for non-critical stuff
Fixes https://github.com/TinyWebEx/RandomTips/issues/7

The issue is here that a promise is usually expected, so I cannot just return
Promise.ject() with an empty error, because this also results in a red error
for the promise (with the worse error message "undefined").

Thus I've just made it to return a successful promise or undefined.

I've also changed the similar uncritical error when tips should be shown
according to gobal randomize, but no tip could be shown due to other factors,
so there is no tip to be shown.

Console.trace is BTW well-supported across browsers:
https://developer.mozilla.org/en-US/docs/Web/API/console/trace#browser_compatibility